third_party/blink/web_tests/fast/css3-text/css3-text-align-last/getComputedStyle/getComputedStyle-text-align-last-expected.txt

This test checks that text-align-last parses properly the properties from CSS 3 Text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test the initial value:
PASS computedStyle.textAlignLast is 'auto'


Value 'start':
PASS e.style.textAlignLast is 'start'
PASS computedStyle.textAlignLast is 'start'


Value 'end':
PASS e.style.textAlignLast is 'end'
PASS computedStyle.textAlignLast is 'end'


Value 'left':
PASS e.style.textAlignLast is 'left'
PASS computedStyle.textAlignLast is 'left'


Value 'right':
PASS e.style.textAlignLast is 'right'
PASS computedStyle.textAlignLast is 'right'


Value 'center':
PASS e.style.textAlignLast is 'center'
PASS computedStyle.textAlignLast is 'center'


Value 'justify':
PASS e.style.textAlignLast is 'justify'
PASS computedStyle.textAlignLast is 'justify'


Value 'auto':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


Invalid value test - '-webkit-left':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


Invalid value test - '-webkit-right':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


Invalid value test - '-webkit-center':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


Invalid value test - '-webkit-match-parent':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


Invalid value test - '-webkit-auto':
PASS e.style.textAlignLast is 'auto'
PASS computedStyle.textAlignLast is 'auto'


PASS successfullyParsed is true

TEST COMPLETE

hello world
